(refactor): 토큰발급 resource 수정 및 네이밍 변경
