+++ app/api/auth/resource/request.go
@@ -1,0 +1,7 @@
+package resource
+
+type CreateTokenRequest struct {
+	Id            uint64 `json:"id"`
+	Provider      string `json:"provider"`
+	FirebaseToken string `json:"firebase_token"`
+}
