+++ app/api/auth/resource/response.go
@@ -1,0 +1,7 @@
+package resource
+
+type CreateTokenResponse struct {
+	AccessToken  string `json:"access_token"`
+	Expired      int    `json:"expired"`
+	RefreshToken string `json:"refresh_token"`
+}
