+++ database/repository/auth/auth.go
@@ -1,4 +1,4 @@
-package token
+package auth
 
 import (
 	"main/database/entity"
